add temp resize listener
